feat(icon): require icon a11y-label or a11y-hidden
The `a11y-label` or `a11y-label` attribute is now required
on the `pharos-icon` component. This is a breaking change
because it will cause a runtime error if one of the
attribute is not provided.

The description attribute will temporarily be allowed as an
alternative to a11y-label until the next major release. A console
warning will let users know that the description attribute has been
deprecated.